+++ packages/cli/src/commands/dev/index.ts
@@ -1,6 +1,6 @@
 import path from 'path';
 import chalk from 'chalk';
-import { PackageJson } from '@vercel/build-utils';
+import { PackageJson } from '@khulnasoft/build-utils';
 
 import { parseArguments } from '../../util/get-args';
 import getSubcommand from '../../util/get-subcommand';