+++ packages/redwood/src/index.ts
@@ -33,7 +33,7 @@ import {
   NodejsLambda,
   BuildV2,
   PrepareCache,
-} from '@vercel/build-utils';
+} from '@khulnasoft/build-utils';
 import { nodeFileTrace } from '@vercel/nft';
 import { getTransformedRoutes, Route } from '@vercel/routing-utils';
 
@@ -108,7 +108,9 @@ export const build: BuildV2 = async ({
   }
 
   if (meta.isDev) {
-    throw new Error('Detected `@vercel/redwood` dev but this is not supported');
+    throw new Error(
+      'Detected `@khulnasoft/redwood` dev but this is not supported'
+    );
   }
 
   const pkg = await readConfigFile<PackageJson>(join(workPath, 'package.json'));