+++ packages/static-build/src/utils/hugo.ts
@@ -1,5 +1,5 @@
 import fetch from 'node-fetch';
-import { NowBuildError } from '@vercel/build-utils';
+import { NowBuildError } from '@khulnasoft/build-utils';
 
 export async function getHugoUrl(
   version: string,