+++ packages/redwood/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @vercel/redwood
+# @khulnasoft/redwood
 
 ## 2.1.3